adicionado tratamento de exceções

+++ ClientService/src/main/java/backend/ClientMicroservice/service/RegisterClient.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import backend.ClientMicroservice.domain.Client;
+import backend.ClientMicroservice.exception.DomainEntityNotFound;
 import backend.ClientMicroservice.repository.IClientRepository;
 import jakarta.validation.Valid;
 
@@ -21,10 +22,16 @@ public class RegisterClient {
     }
 
     public Client updateClient(@Valid Client client) {
+        if (!clientRepository.existsById(client.getId())) {
+            throw new DomainEntityNotFound(Client.class, "id", client.getId());
+        }
         return this.clientRepository.save(client);
     }
 
     public void deleteClient(String id) {
+        if (!clientRepository.existsById(id)) {
+            throw new DomainEntityNotFound(Client.class, "id", id);
+        }
         this.clientRepository.deleteById(id);
     }
 }
