+++ ProductService/src/main/java/backend/ProductMicroservice/service/RegisterProduct.java
@@ -4,6 +4,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import backend.ProductMicroservice.domain.Product;
+import backend.ProductMicroservice.exception.DomainEntityNotFound;
 import backend.ProductMicroservice.repository.IProductRepository;
 import jakarta.validation.Valid;
 
@@ -21,10 +22,16 @@ public class RegisterProduct {
     }
 
     public Product updateProduct(@Valid Product product) {
+        if (!productRepository.findById(product.getId()).isPresent()) {
+            throw new DomainEntityNotFound(Product.class,"id", product.getId());
+        }
         return this.productRepository.save(product);
     }
 
     public void deleteProduct(String id) {
+        if (!productRepository.findById(id).isPresent()) {
+            throw new DomainEntityNotFound(Product.class,"id", id);
+        }
         this.productRepository.deleteById(id);
     }
 }
